Remote IP address for node upstream of nginx
Edited the way the remote IP Address is found when node is being
proxied/upstreamed through a nginx instance. It checks the
header['x-forwarded-for'] first to see if a upstream IP exists. If not,
it falls back to socket.remoteAddress for the IP.

Also added a console.log when a IP has reached the rate limit.